requirements.txt
beautifulsoup4
databases[postgresql]
requests
python-environ